Refactor zsh init scripts

+++ zsh/tmux.zsh
@@ -1,4 +1,6 @@
-alias ta='tmux attach-session -t'
-alias tk='tmux kill-session -t'
-alias tl='tmux list-sessions'
-alias tn='tmux new-session -s'
+if type tmux > /dev/null; then
+  alias ta='tmux attach-session -t'
+  alias tk='tmux kill-session -t'
+  alias tl='tmux list-sessions'
+  alias tn='tmux new-session -s'
+fi
